Write ~(unsigned int)_MM_ROUND_MASK where the headers clear the rounding bits of MXCSR, whose negative int GCC warned about with -Wsign-conversion in every file including them, and pass the exponent of nth_root() and the mask of the tests as unsigned
_MM_ROUND_MASK is the int 0x6000: ~_MM_ROUND_MASK is the int -24577, which
_mm_getcsr() & ~_MM_ROUND_MASK converts to unsigned int. GCC 9.4 with
-Wsign-conversion warned about it at gaol_fpu_fenv.h:121, in the function
writing the control registers the previous commits added, and at
gaol_fpu.h:142, in set_rounding(): 14 times each in GAOL's build, and in every
file of a program including GAOL's headers. Clang 18, whose -Wconversion
includes -Wsign-conversion, did not warn. The mask is now made unsigned before
it is complemented, which gives the same bits; the same for the Visual C++
version of the function and for GAOL_RND_LEAVE_SSE().

In the tests, arithmetic.cpp passed an int to nth_root(), which takes an
unsigned int, and rounding_direction.cpp complemented the same mask.

GCC 9.4 with -Wall -Wconversion -Wsign-conversion and Clang 18 with -Wall
-Wconversion compile GAOL and its tests without warnings; the tests pass.

+++ gaol/gaol_fpu.h
@@ -87,7 +87,7 @@
 #  define GAOL_RND_KEEP(x)      ((x) = gaol::rnd_keep(x))
 #  if USING_SSE2_INSTRUCTIONS
 #     define GAOL_RND_ENTER_SSE() const unsigned int _save_state_sse = _mm_getcsr() & _MM_ROUND_MASK; gaol::round_upward_sse()
-#     define GAOL_RND_LEAVE_SSE() _mm_setcsr((_mm_getcsr() & ~_MM_ROUND_MASK) | _save_state_sse)
+#     define GAOL_RND_LEAVE_SSE() _mm_setcsr((_mm_getcsr() & ~(unsigned int)_MM_ROUND_MASK) | _save_state_sse)
 #  endif
 #else // !GAOL_PRESERVE_ROUNDING
 #  define GAOL_RND_ENTER()      gaol::round_upward_if_needed()
@@ -139,7 +139,7 @@ namespace gaol {
   {
     fesetround(s.direction);
 #if GAOL_RND_SSE_REGISTER
-    _mm_setcsr((_mm_getcsr() & ~_MM_ROUND_MASK) | s.sse);
+    _mm_setcsr((_mm_getcsr() & ~(unsigned int)_MM_ROUND_MASK) | s.sse);
 #endif
   }
